merge-octopus: avoid dashed invocation
A long time ago, in turn a long time after introducing builtins and
moving Git's scripts to libexec/git-core/, we deprecated the invocation
of dashed commands.

The merge-octopus script is a holdover from the time before that, as it
still tries to invoke a builtin by its dashed form. Let's just not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>